apps/ai-engine/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
readability-lxml
redis
supabase
python-dotenv
openai 